@@ -1,50 +1,50 @@
 name: inventree
 description: InvenTree stock management
 
-version: 0.9.0+50
+version: 0.10.2+57
 
 environment:
   sdk: ">=2.16.0 <3.0.0"
 
 dependencies:
-
-  audioplayers: ^0.20.1                   # Play audio files
+  audioplayers: ^3.0.1                    # Play audio files
   cached_network_image: ^3.2.0            # Download and cache remote images
-  camera: ^0.9.4                          # Camera
+  camera: ^0.10.3                         # Camera
   cupertino_icons: ^1.0.3
+  currency_formatter: ^2.0.0
   datetime_picker_formfield: ^2.0.0       # Date / time picker
-  device_info_plus: ^3.2.2                # Information about the device
-  dropdown_search: ^0.6.3                 # Dropdown autocomplete form fields
-  file_picker: ^4.5.1                     # Select files from the device
+  device_info_plus: ^8.0.0                # Information about the device
+  dropdown_search: ^5.0.5                 # Dropdown autocomplete form fields
+  file_picker: ^5.2.5                     # Select files from the device
   flutter:
     sdk: flutter
   flutter_cache_manager: ^3.3.0
   flutter_localizations:
     sdk: flutter
-  flutter_localized_locales: 2.0.3
-  flutter_markdown: ^0.6.9                # Rendering markdown
+  flutter_localized_locales: ^2.0.4
+  flutter_markdown: ^0.6.13+1             # Rendering markdown
   flutter_overlay_loader: ^2.0.0          # Overlay screen support
-  font_awesome_flutter: ^9.1.0            # FontAwesome icon set
+  font_awesome_flutter: ^10.3.0           # FontAwesome icon set
   http: ^0.13.4
-  image_picker: ^0.8.3                    # Select or take photos
+  image_picker: ^0.8.6+1                  # Select or take photos
   infinite_scroll_pagination: ^3.1.0      # Let the server do all the work!
   intl: ^0.17.0
-  one_context: ^1.1.1                     # Dialogs without requiring context
-  open_file: ^3.2.1                       # Open local files
-  package_info_plus: ^1.0.4               # App information introspection
-  path: ^1.8.0
-  path_provider: ^2.0.2                   # Local file storage
-  qr_code_scanner: ^0.7.0                 # Barcode scanning
-  sembast: ^3.1.0+2                       # NoSQL data storage
-  sentry_flutter: ^6.4.0                  # Error reporting
-  url_launcher: ^6.0.9                    # Open link in system browser
+  one_context: ^2.1.0                     # Dialogs without requiring context
+  open_filex: ^4.3.2                      # Open local files
+  package_info_plus: ^3.0.2               # App information introspection
+  path: ^1.8.2
+  path_provider: ^2.0.12                  # Local file storage
+  qr_code_scanner: ^1.0.1                 # Barcode scanning
+  sembast: ^3.4.0+6                       # NoSQL data storage
+  sentry_flutter: ^6.19.0                 # Error reporting
+  url_launcher: ^6.1.9                    # Open link in system browser
 
 dev_dependencies:
-  flutter_launcher_icons: ^0.9.0
+  flutter_launcher_icons: ^0.11.0
   flutter_test:
     sdk: flutter
-  lint: ^1.8.0
-  test: ^1.19.0
+  lint: ^2.0.1
+  test: ^1.22.0
 
 flutter_icons:
   android: true
