@@ -10,8 +10,9 @@ import "package:url_launcher/url_launcher.dart";
 
 import "package:path/path.dart" as path;
 
-import "package:inventree/l10.dart";
 import "package:inventree/api_form.dart";
+import "package:inventree/fa_icon_mapping.dart";
+import "package:inventree/l10.dart";
 
 
 // Paginated response object
@@ -148,6 +149,61 @@ class InvenTreeModel {
   // Legacy API provided external link as "URL", while newer API uses "link"
   String get link => (jsondata["link"] ?? jsondata["URL"] ?? "") as String;
 
+  /*
+   * Attempt to extract a custom icon for this model.
+   * If icon data is provided, attempt to convert to a FontAwesome icon
+   *
+   * Icon data *should* be presented something like "fas fa-boxes" / "fab fa-github" (etc):
+   *
+   * - First part specifies the *style*
+   * - Second part specifies the icon
+   *
+   */
+  FaIcon? get customIcon {
+    String icon = (jsondata["icon"] ?? "").toString();
+
+    // Empty icon (default)
+    if (icon.isEmpty) {
+      return null;
+    }
+
+    final split = icon.trim().split(" ");
+
+    // Must have two distinct units
+    if (split.length != 2) {
+      return null;
+    }
+
+    String style = split[0];
+    String name = split[1];
+
+    // Remove "fa-" leading text (if provided)
+    if (name.startsWith("fa-")) {
+      name = name.substring(3);
+    }
+
+    int iconHex = fontAwesomeIconMap[name] ?? 0;
+
+    // No match for the icon name
+    if (iconHex == 0) {
+      return null;
+    }
+
+    switch (style) {
+      case "fas":
+        return FaIcon(IconDataSolid(iconHex));
+      case "fab":
+        return FaIcon(IconDataBrands(iconHex));
+      case "fa":
+        return FaIcon(IconDataRegular(iconHex));
+      case "fal":
+        return FaIcon(IconDataLight(iconHex));
+      default:
+        // No match
+        return null;
+    }
+  }
+
   /* Extract any custom barcode data available for the model.
    * Note that old API used 'uid' (only for StockItem),
    * but this was updated to use 'barcode_hash'
@@ -167,8 +223,9 @@ class InvenTreeModel {
 
   Future <void> goToInvenTreePage() async {
 
-    if (await canLaunch(webUrl)) {
-      await launch(webUrl);
+    var uri = Uri.tryParse(webUrl);
+    if (uri != null && await canLaunchUrl(uri)) {
+      await launchUrl(uri);
     } else {
       // TODO
     }
@@ -177,9 +234,9 @@ class InvenTreeModel {
   Future <void> openLink() async {
 
     if (link.isNotEmpty) {
-
-      if (await canLaunch(link)) {
-        await launch(link);
+      var uri = Uri.tryParse(link);
+      if (uri != null && await canLaunchUrl(uri)) {
+        await launchUrl(uri);
       }
     }
   }
@@ -623,7 +680,19 @@ class InvenTreePlugin extends InvenTreeModel {
   }
 
   @override
-  String get URL => "plugin/";
+  String get URL {
+
+    /* Note: The plugin API endpoint changed at API version 90,
+     * <  90 = 'plugin'
+     * >= 90 = 'plugins'
+     * Ref: https://github.com/inventree/InvenTree/pull/4186
+     */
+    if (api.isConnected() && api.apiVersion < 90) {
+      return "plugin/";
+    } else {
+      return "plugins/";
+    }
+  }
 
   String get key => (jsondata["key"] ?? "") as String;
 
@@ -731,7 +800,7 @@ class InvenTreeAttachment extends InvenTreeModel {
       }
     }
 
-    return FontAwesomeIcons.fileAlt;
+    return FontAwesomeIcons.fileLines;
   }
 
   String get comment => (jsondata["comment"] ?? "") as String;
