@@ -8,6 +8,7 @@
  */
 
 import "dart:io";
+import "package:currency_formatter/currency_formatter.dart";
 
 import "package:audioplayers/audioplayers.dart";
 import "package:one_context/one_context.dart";
@@ -74,6 +75,37 @@ Future<void> playAudioFile(String path) async {
     return;
   }
 
-  final player = AudioCache();
-  player.play(path);
+  final player = AudioPlayer();
+  player.play(AssetSource(path));
+}
+
+
+/*
+ * Helper function for rendering a money / currency object as a String
+ */
+String renderCurrency(double? amount, String currency, {int decimals = 2}) {
+
+  if (amount == null) return "-";
+  if (amount.isInfinite || amount.isNaN) return "-";
+
+  currency = currency.trim();
+
+  if (currency.isEmpty) return "-";
+
+  CurrencyFormatterSettings backupSettings = CurrencyFormatterSettings(
+    symbol: "\$",
+    symbolSide: SymbolSide.left,
+  );
+
+  String value = CurrencyFormatter.format(
+    amount,
+    CurrencyFormatter.majors[currency.toLowerCase()] ?? backupSettings
+  );
+
+  // If we were not able to determine the currency
+  if (!CurrencyFormatter.majors.containsKey(currency.toLowerCase())) {
+    value += " ${currency}";
+  }
+
+  return value;
 }
