@@ -87,6 +87,26 @@ class InvenTreeCompany extends InvenTreeModel {
 
 
 /*
+ * Class representing an attachment file against a Company object
+ */
+class InvenTreeCompanyAttachment extends InvenTreeAttachment {
+
+  InvenTreeCompanyAttachment() : super();
+
+  InvenTreeCompanyAttachment.fromJson(Map<String, dynamic> json) : super.fromJson(json);
+
+  @override
+  String get REFERENCE_FIELD => "company";
+
+  @override
+  String get URL => "company/attachment/";
+
+  @override
+  InvenTreeModel createFromJson(Map<String, dynamic> json) => InvenTreeCompanyAttachment.fromJson(json);
+
+}
+
+/*
  * The InvenTreeSupplierPart class represents the SupplierPart model in the InvenTree database
  */
 class InvenTreeSupplierPart extends InvenTreeModel {
@@ -98,11 +118,23 @@ class InvenTreeSupplierPart extends InvenTreeModel {
   @override
   String get URL => "company/part/";
 
+  @override
+  Map<String, dynamic> formFields() {
+    return {
+      "supplier": {},
+      "SKU": {},
+      "link": {},
+      "note": {},
+      "packaging": {},
+      "pack_size": {},
+    };
+  }
+
   Map<String, String> _filters() {
     return {
       "manufacturer_detail": "true",
       "supplier_detail": "true",
-      "manufacturer_part_detail": "true",
+      "part_detail": "true",
     };
   }
 
@@ -116,29 +148,31 @@ class InvenTreeSupplierPart extends InvenTreeModel {
     return _filters();
   }
 
-  int get manufacturerId => (jsondata["manufacturer"] ?? -1) as int;
+  String get manufacturerName => (jsondata["manufacturer_detail"]?["name"] ?? "") as String;
 
-  String get manufacturerName => (jsondata["manufacturer_detail"]["name"] ?? "") as String;
+  String get MPN => (jsondata["manufacturer_part_detail"]?["MPN"] ?? "") as String;
 
-  String get manufacturerImage => (jsondata["manufacturer_detail"]["image"] ?? jsondata["manufacturer_detail"]["thumbnail"] ?? InvenTreeAPI.staticThumb) as String;
+  String get manufacturerImage => (jsondata["manufacturer_detail"]?["image"] ?? jsondata["manufacturer_detail"]["thumbnail"] ?? InvenTreeAPI.staticThumb) as String;
 
   int get manufacturerPartId => (jsondata["manufacturer_part"] ?? -1) as int;
 
   int get supplierId => (jsondata["supplier"] ?? -1) as int;
 
-  String get supplierName => (jsondata["supplier_detail"]["name"] ?? "") as String;
+  String get supplierName => (jsondata["supplier_detail"]?["name"] ?? "") as String;
 
-  String get supplierImage => (jsondata["supplier_detail"]["image"] ?? jsondata["supplier_detail"]["thumbnail"] ?? InvenTreeAPI.staticThumb) as String;
+  String get supplierImage => (jsondata["supplier_detail"]?["image"] ?? jsondata["supplier_detail"]["thumbnail"] ?? InvenTreeAPI.staticThumb) as String;
 
   String get SKU => (jsondata["SKU"] ?? "") as String;
 
-  String get MPN => (jsondata["MPN"] ?? "") as String;
-
   int get partId => (jsondata["part"] ?? -1) as int;
 
-  String get partImage => (jsondata["part_detail"]["thumbnail"] ?? InvenTreeAPI.staticThumb) as String;
+  String get partImage => (jsondata["part_detail"]?["thumbnail"] ?? InvenTreeAPI.staticThumb) as String;
 
-  String get partName => (jsondata["part_detail"]["full_name"] ?? "") as String;
+  String get partName => (jsondata["part_detail"]?["full_name"] ?? "") as String;
+
+  String get partDescription => (jsondata["part_detail"]?["description"] ?? "") as String;
+
+  String get note => (jsondata["note"] ?? "") as String;
 
   @override
   InvenTreeModel createFromJson(Map<String, dynamic> json) {
