@@ -1,6 +1,7 @@
 import "package:flutter/material.dart";
 import "package:flutter_markdown/flutter_markdown.dart";
 import "package:inventree/l10.dart";
+import "package:url_launcher/url_launcher.dart";
 
 
 class ReleaseNotesWidget extends StatelessWidget {
@@ -30,6 +31,18 @@ class CreditsWidget extends StatelessWidget {
 
   final String credits;
 
+  /*
+   * Callback function when a link is clicked in the markdown
+   */
+  Future<void> openLink(String url) async {
+
+    final link = Uri.parse(url);
+
+    if (await canLaunchUrl(link)) {
+      await launchUrl(link);
+    }
+  }
+
   @override
   Widget build (BuildContext context) {
     return Scaffold(
@@ -38,7 +51,13 @@ class CreditsWidget extends StatelessWidget {
       ),
       body: Markdown(
         selectable: false,
-        data: credits
+        data: credits,
+        onTapLink: (url, href, title) {
+          var link = href ?? "";
+          if (link.isNotEmpty) {
+            openLink(link);
+          }
+        },
       )
     );
   }
