@@ -1,14 +1,24 @@
-import "package:inventree/api.dart";
-import "package:inventree/barcode.dart";
 import "package:flutter/material.dart";
-import "package:inventree/l10.dart";
-import "package:inventree/settings/about.dart";
-
-import "package:inventree/settings/settings.dart";
 import "package:font_awesome_flutter/font_awesome_flutter.dart";
-import "package:inventree/widget/search.dart";
 import "package:package_info_plus/package_info_plus.dart";
 
+import "package:inventree/api.dart";
+import "package:inventree/barcode.dart";
+import "package:inventree/l10.dart";
+
+import "package:inventree/settings/about.dart";
+import "package:inventree/settings/settings.dart";
+
+import "package:inventree/widget/search.dart";
+
+
+/*
+ * Custom "drawer" widget for the InvenTree app.
+ *
+ * - Provides a "home" button which completely unwinds the widget stack
+ * - Global search
+ * - Barcoed scan
+ */
 class InvenTreeDrawer extends StatelessWidget {
 
   const InvenTreeDrawer(this.context);
@@ -86,7 +96,7 @@ class InvenTreeDrawer extends StatelessWidget {
               context: context,
               tiles: <Widget>[
                 ListTile(
-                  leading: FaIcon(FontAwesomeIcons.home),
+                  leading: FaIcon(FontAwesomeIcons.house),
                   title: Text(
                     L10().appTitle,
                     style: TextStyle(fontWeight: FontWeight.bold),
@@ -100,7 +110,7 @@ class InvenTreeDrawer extends StatelessWidget {
                 ),
                 ListTile(
                   title: Text(L10().search),
-                  leading: FaIcon(FontAwesomeIcons.search),
+                  leading: FaIcon(FontAwesomeIcons.magnifyingGlass),
                   onTap: _search,
                 ),
                 ListTile(
@@ -110,7 +120,7 @@ class InvenTreeDrawer extends StatelessWidget {
                 ),
                 ListTile(
                   title: Text(L10().about),
-                  leading: FaIcon(FontAwesomeIcons.infoCircle),
+                  leading: FaIcon(FontAwesomeIcons.circleInfo),
                   onTap: _about,
                 )
               ]
